Projects/01_Handwritten Digit Recognition/requirements.txt
# Core
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# Deep Learning
tensorflow==2.13.0
keras==2.13.1

# Utilities
scikit-learn==1.3.0
opencv-python==4.8.1.78
pillow==10.0.1
tqdm==4.65.0

# Deployment
flask==2.3.3
streamlit==1.27.0
gunicorn==21.2.0

# Development
jupyter==1.0.0
ipykernel==6.25.1
black==23.7.0
pytest==7.4.0